功能完善

+++ app/src/main/java/info/free/duangjike/ThemeUtil.kt
@@ -12,12 +12,12 @@ object ThemeUtil {
     /**
      * 根据手机的分辨率从 dp 的单位 转成为 px(像素)
      */
-    fun dip2px(dpValue: Int): Int {
+    fun dip2px(dpValue: Int): Float {
         val context = DuangApplication.context
         if (scale == -1f) {
             scale = context.resources.displayMetrics.density
         }
-        return (dpValue * scale + 0.5f).toInt()
+        return (dpValue * scale + 0.5f)
     }
     /**
      *
